Add files via upload
This project uses a weighted graph to find a set of related tags associated with a certain tag based on the frequency with which the queried tag is associated with other tags. 
The tags are extracted from the StackExchange website through a private API key. Please request a private key(to make unlimited api calls). Please refer to the stackExchange API documentation